# docs/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Twitter API
  description: |
  version: 1.0.0
  contact:
    name: API Support

servers:
  - url: /api
    description: Main API server

paths:
  /users/{username}:
    get:
      summary: Get user profile
      description: Fetches detailed profile information for a specific Twitter user using FxTwitter API
      operationId: getUser
      tags:
        - Users
      parameters:
        - name: username
          in: path
          required: true
          description: Twitter username (without @)
          schema:
            type: string
            example: elonmusk
      responses:
        '200':
          description: Successfully retrieved user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FxTwitterUserResponse'
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
              example: "Failed to fetch user: connection timeout"

  /users/{username}/tweets:
    get:
      summary: Get user tweets
      description: Fetches a list of tweet IDs for a specific Twitter user from Nitter RSS feed
      operationId: getUserTweets
      tags:
        - Tweets
      parameters:
        - name: username
          in: path
          required: true
          description: Twitter username (without @)
          schema:
            type: string
            example: elonmusk
      responses:
        '200':
          description: Successfully retrieved tweet IDs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TweetsResponse'
              example:
                username: "elonmusk"
                tweet_ids:
                  - "1234567890123456789"
                  - "1234567890123456790"
                  - "1234567890123456791"
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
              example: "Failed to fetch tweets: connection timeout"

  /users/{username}/tweets/{id}:
    get:
      summary: Get tweet details
      description: Fetches detailed information about a specific tweet using FxTwitter API
      operationId: getTweet
      tags:
        - Tweets
      parameters:
        - name: username
          in: path
          required: true
          description: Twitter username (without @)
          schema:
            type: string
            example: elonmusk
        - name: id
          in: path
          required: true
          description: Tweet ID
          schema:
            type: string
            example: "1234567890123456789"
      responses:
        '200':
          description: Successfully retrieved tweet details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FxTwitterResponse'
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
              example: "Failed to fetch tweet: connection timeout"

components:
  schemas:
    TweetsResponse:
      type: object
      required:
        - username
        - tweet_ids
      properties:
        username:
          type: string
          description: Twitter username
          example: "elonmusk"
        tweet_ids:
          type: array
          description: List of tweet IDs
          items:
            type: string
          example:
            - "1234567890123456789"
            - "1234567890123456790"

    FxTwitterResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: HTTP status code
          example: 200
        message:
          type: string
          description: Response message
          example: "OK"
        tweet:
          $ref: '#/components/schemas/Tweet'

    Tweet:
      type: object
      required:
        - url
        - id
        - text
        - author
        - replies
        - retweets
        - likes
        - created_at
        - created_timestamp
        - possibly_scam
        - possibly_sensitive
        - lang
        - source
      properties:
        url:
          type: string
          format: uri
          description: Tweet URL
          example: "https://twitter.com/elonmusk/status/1234567890123456789"
        id:
          type: string
          description: Tweet ID
          example: "1234567890123456789"
        text:
          type: string
          description: Tweet text content
          example: "Hello Twitter!"
        author:
          $ref: '#/components/schemas/Author'
        replies:
          type: integer
          format: int64
          description: Number of replies
          example: 42
        retweets:
          type: integer
          format: int64
          description: Number of retweets
          example: 150
        likes:
          type: integer
          format: int64
          description: Number of likes
          example: 1000
        views:
          type: integer
          format: int64
          nullable: true
          description: Number of views (if available)
          example: 50000
        created_at:
          type: string
          format: date-time
          description: Tweet creation timestamp (ISO 8601)
          example: "2024-01-15T12:30:00Z"
        created_timestamp:
          type: integer
          format: int64
          description: Unix timestamp of tweet creation
          example: 1705324200
        possibly_scam:
          type: boolean
          description: Flag indicating if tweet might be a scam
          example: false
        possibly_sensitive:
          type: boolean
          description: Flag indicating if tweet contains sensitive content
          example: false
        lang:
          type: string
          description: Tweet language code
          example: "en"
        source:
          type: string
          description: Client used to post the tweet
          example: "Twitter for iPhone"
        media:
          $ref: '#/components/schemas/Media'
        poll:
          $ref: '#/components/schemas/Poll'
        quote:
          $ref: '#/components/schemas/Tweet'
        translation:
          $ref: '#/components/schemas/Translation'

    Author:
      type: object
      required:
        - id
        - name
        - screen_name
        - avatar_url
        - verified
        - blue_badge
      properties:
        id:
          type: string
          description: Author's user ID
          example: "44196397"
        name:
          type: string
          description: Author's display name
          example: "Elon Musk"
        screen_name:
          type: string
          description: Author's username
          example: "elonmusk"
        avatar_url:
          type: string
          format: uri
          description: URL to author's avatar image
          example: "https://pbs.twimg.com/profile_images/1234567890/avatar.jpg"
        avatar_color:
          type: string
          nullable: true
          description: Hex color of avatar background
          example: "#1DA1F2"
        banner_url:
          type: string
          format: uri
          nullable: true
          description: URL to author's banner image
          example: "https://pbs.twimg.com/profile_banners/44196397/1234567890"
        description:
          type: string
          nullable: true
          description: Author's bio
          example: "Tesla, SpaceX, etc."
        location:
          type: string
          nullable: true
          description: Author's location
          example: "San Francisco, CA"
        url:
          type: string
          format: uri
          nullable: true
          description: Author's website URL
          example: "https://tesla.com"
        followers:
          type: integer
          format: int64
          nullable: true
          description: Number of followers
          example: 100000000
        following:
          type: integer
          format: int64
          nullable: true
          description: Number of accounts following
          example: 100
        joined:
          type: string
          nullable: true
          description: Account creation date
          example: "June 2009"
        likes:
          type: integer
          format: int64
          nullable: true
          description: Number of likes given
          example: 50000
        tweets:
          type: integer
          format: int64
          nullable: true
          description: Total number of tweets
          example: 25000
        verified:
          type: boolean
          description: Legacy verified badge status
          example: true
        blue_badge:
          type: boolean
          description: Twitter Blue subscription status
          example: true

    Media:
      type: object
      properties:
        all:
          type: array
          description: All media items
          items:
            $ref: '#/components/schemas/MediaItem'
        photos:
          type: array
          description: Photo attachments
          items:
            $ref: '#/components/schemas/Photo'
        videos:
          type: array
          description: Video attachments
          items:
            $ref: '#/components/schemas/Video'
        mosaic:
          $ref: '#/components/schemas/MosaicInfo'
        external:
          $ref: '#/components/schemas/ExternalMedia'

    MediaItem:
      type: object
      required:
        - type
        - url
        - width
        - height
      properties:
        type:
          type: string
          description: Media type
          enum: [photo, video, gif]
          example: "photo"
        url:
          type: string
          format: uri
          description: Media URL
          example: "https://pbs.twimg.com/media/image.jpg"
        width:
          type: integer
          description: Media width in pixels
          example: 1200
        height:
          type: integer
          description: Media height in pixels
          example: 800
        format:
          type: string
          description: File format
          example: "jpg"
        thumbnail_url:
          type: string
          format: uri
          description: Thumbnail URL (for videos)
          example: "https://pbs.twimg.com/media/thumb.jpg"
        duration:
          type: integer
          nullable: true
          description: Duration in seconds (for videos)
          example: 30

    Photo:
      type: object
      required:
        - url
        - width
        - height
      properties:
        url:
          type: string
          format: uri
          description: Photo URL
          example: "https://pbs.twimg.com/media/image.jpg"
        width:
          type: integer
          description: Photo width in pixels
          example: 1200
        height:
          type: integer
          description: Photo height in pixels
          example: 800
        format:
          type: string
          description: Image format
          example: "jpg"

    Video:
      type: object
      required:
        - url
        - thumbnail_url
        - width
        - height
      properties:
        url:
          type: string
          format: uri
          description: Video URL
          example: "https://video.twimg.com/ext_tw_video/video.mp4"
        thumbnail_url:
          type: string
          format: uri
          description: Video thumbnail URL
          example: "https://pbs.twimg.com/media/thumb.jpg"
        width:
          type: integer
          description: Video width in pixels
          example: 1280
        height:
          type: integer
          description: Video height in pixels
          example: 720
        format:
          type: string
          description: Video format
          example: "mp4"
        duration:
          type: integer
          description: Duration in seconds
          example: 45
        type:
          type: string
          description: Video type
          example: "video"

    MosaicInfo:
      type: object
      required:
        - type
      properties:
        type:
          type: string
          description: Mosaic layout type
          example: "mosaic_2x2"
        formats:
          type: array
          items:
            type: string
          description: Available formats
          example: ["jpg", "webp"]

    ExternalMedia:
      type: object
      required:
        - type
        - url
      properties:
        type:
          type: string
          description: External media type
          example: "youtube"
        url:
          type: string
          format: uri
          description: External media URL
          example: "https://www.youtube.com/watch?v=dQw4w9WgXcQ"
        thumbnail_url:
          type: string
          format: uri
          description: Thumbnail URL
          example: "https://i.ytimg.com/vi/dQw4w9WgXcQ/maxresdefault.jpg"
        width:
          type: integer
          description: Width in pixels
          example: 1280
        height:
          type: integer
          description: Height in pixels
          example: 720
        duration:
          type: integer
          nullable: true
          description: Duration in seconds
          example: 212

    Poll:
      type: object
      required:
        - total_votes
        - ends_at
        - time_remaining
        - choices
      properties:
        total_votes:
          type: integer
          format: int64
          description: Total number of votes
          example: 10000
        ends_at:
          type: string
          format: date-time
          description: Poll end timestamp
          example: "2024-01-20T12:00:00Z"
        time_remaining:
          type: string
          description: Human-readable time remaining
          example: "2 days left"
        choices:
          type: array
          description: Poll choices
          items:
            $ref: '#/components/schemas/PollChoice'

    PollChoice:
      type: object
      required:
        - label
        - count
        - percentage
      properties:
        label:
          type: string
          description: Choice label
          example: "Yes"
        count:
          type: integer
          format: int64
          description: Number of votes for this choice
          example: 7500
        percentage:
          type: integer
          description: Percentage of total votes
          example: 75

    Translation:
      type: object
      required:
        - text
        - source_lang
        - target_lang
      properties:
        text:
          type: string
          description: Translated text
          example: "Hello Twitter!"
        source_lang:
          type: string
          description: Source language code
          example: "es"
        target_lang:
          type: string
          description: Target language code
          example: "en"
        translation_url:
          type: string
          format: uri
          description: URL to translation service
          example: "https://translate.twitter.com/..."

    FxTwitterUserResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: HTTP status code
          example: 200
        message:
          type: string
          description: Response message
          example: "OK"
        user:
          $ref: '#/components/schemas/User'

    User:
      type: object
      required:
        - screen_name
        - url
        - id
        - followers
        - following
        - likes
        - media_count
        - tweets
        - name
        - description
        - location
        - banner_url
        - avatar_url
        - joined
        - protected
      properties:
        screen_name:
          type: string
          description: User's Twitter handle
          example: "elonmusk"
        url:
          type: string
          format: uri
          description: URL to user's Twitter profile
          example: "https://x.com/elonmusk"
        id:
          type: string
          description: User's unique ID
          example: "44196397"
        followers:
          type: integer
          format: int64
          description: Number of followers
          example: 228344297
        following:
          type: integer
          format: int64
          description: Number of accounts following
          example: 1224
        likes:
          type: integer
          format: int64
          description: Total number of likes given
          example: 179542
        media_count:
          type: integer
          format: int64
          description: Number of media items posted
          example: 4200
        tweets:
          type: integer
          format: int64
          description: Total number of tweets
          example: 88225
        name:
          type: string
          description: User's display name
          example: "Elon Musk"
        description:
          type: string
          description: User's bio/description
          example: ""
        location:
          type: string
          description: User's location
          example: ""
        banner_url:
          type: string
          format: uri
          description: URL to user's banner image
          example: "https://pbs.twimg.com/profile_banners/44196397/1739948056"
        avatar_url:
          type: string
          format: uri
          description: URL to user's profile picture
          example: "https://pbs.twimg.com/profile_images/1936002956333080576/kqqe2iWO_normal.jpg"
        joined:
          type: string
          description: Account creation date
          example: "Tue Jun 02 20:12:29 +0000 2009"
        protected:
          type: boolean
          description: Whether the account is protected (private)
          example: false
        website:
          type: string
          format: uri
          nullable: true
          description: User's website URL
          example: null
        verification:
          $ref: '#/components/schemas/Verification'

    Verification:
      type: object
      required:
        - verified
        - type
      properties:
        verified:
          type: boolean
          description: Whether the user is verified
          example: true
        type:
          type: string
          description: Type of verification
          enum: [individual, business, government]
          example: "individual"

tags:
  - name: Users
    description: Operations for fetching user profile data
  - name: Tweets
    description: Operations for fetching tweet data